Modularize writeCsvFile()
